% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{becerra2021}
\alias{becerra2021}
\title{Returns a vector of sample documents}
\usage{
becerra2021()
}
\value{
A character vector.
}
\description{
Returns a vector of sample documents
}
\examples{
docs <- becerra2021()
}
