% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_handlers.R
\name{get_codes}
\alias{get_codes}
\title{Get codes from mincaqdasr data list}
\usage{
get_codes(x)
}
\arguments{
\item{x}{A data list created with mincaqdasr GUI (exported as .json)}
}
\value{
A tibble
}
\description{
Get codes from mincaqdasr data list
}
\examples{
get_codes( mincaqdasr::sample_haitian_fathers_annotated )
}
