% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_handlers.R
\name{get_documents}
\alias{get_documents}
\title{Get documents from mincaqdasr data list}
\usage{
get_documents(x)
}
\arguments{
\item{x}{A data list created with mincaqdasr GUI (exported as .json)}
}
\value{
A tibble
}
\description{
Get documents from mincaqdasr data list
}
\examples{
get_documents( mincaqdasr::sample_big_data_annotated )
}
