% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_handlers.R
\name{get_documents_annotations}
\alias{get_documents_annotations}
\title{Get document annotations from mincaqdasr data list}
\usage{
get_documents_annotations(x)
}
\arguments{
\item{x}{A data list created with mincaqdasr GUI (exported as .json)}
}
\value{
A tibble
}
\description{
Get document annotations from mincaqdasr data list
}
\examples{
get_documents_annotations( mincaqdasr::sample_big_data_annotated )
}
