% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_handlers.R
\name{get_fragments}
\alias{get_fragments}
\title{Get fragments from mincaqdasr data list}
\usage{
get_fragments(x)
}
\arguments{
\item{x}{A data list created with mincaqdasr GUI (exported as .json)}
}
\value{
A tibble
}
\description{
Get fragments from mincaqdasr data list
}
\examples{
get_fragments( mincaqdasr::sample_haitian_fathers_annotated )
}
