% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_handlers.R
\name{get_fragments_annotations}
\alias{get_fragments_annotations}
\title{Get fragment annotations from mincaqdasr data list}
\usage{
get_fragments_annotations(x)
}
\arguments{
\item{x}{A data list created with mincaqdasr GUI (exported as .json)}
}
\value{
A tibble
}
\description{
Get fragment annotations from mincaqdasr data list
}
\examples{
get_fragments_annotations( mincaqdasr::sample_haitian_fathers_annotated )
}
