% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_handlers.R
\name{merge_jsons}
\alias{merge_jsons}
\title{Merge documents, codes and annotations from 2 mincaqdasr data lists}
\usage{
merge_jsons(x, y)
}
\arguments{
\item{x}{A data list created with mincaqdasr GUI (exported as .json)}

\item{y}{A data list created with mincaqdasr GUI (exported as .json)}
}
\value{
A mincaqdasr data list
}
\description{
Merge documents, codes and annotations from 2 mincaqdasr data lists
}
\examples{
a = a
}
