% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_big_data_annotated}
\alias{sample_big_data_annotated}
\title{Fragments of news about big data}
\format{
A list with 4 elements
\describe{
\item{documents}{Character vector with document}
\item{codes}{Characer vector with codes}
\item{documents_annotations}{A data frame with annotation on document level}
\item{documents_annotations$document}{An integer with the position of document in $documents}
\item{documents_annotations$codes}{A list of integers with the position of codes in $codes}
\item{documents_annotations$memo}{A character string with memos}
\item{documents_annotations$annotation_update}{A character string with the timestamp of the last change}
\item{documents_annotations$annotation_user}{A character string with the username}
\item{fragments_annotation}{A data frame with fragments and fragment annotations}
\item{fragments_annotation$id}{A character string with ID hash}
\item{fragments_annotation$document}{An integer with the position of document in $documents}
\item{fragments_annotation$text}{A character string with the fragment's text}
\item{fragments_annotation$start}{An integer with the start position of the fragment in the document}
\item{fragments_annotation$end}{An integer with the en position of the fragment in the document}
\item{fragments_annotation$codes}{A list of integers with the position of codes in $codes}
\item{fragments_annotation$memo}{A character string with memos}
\item{fragments_annotation$annotation_update}{A character string with the timestamp of the last change}
\item{fragments_annotation$annotation_user}{A character string with the username}#'
}
}
\source{
#' https://bookdown.org/gaston_becerra/curso-intro-r/clasificar-automaticamente.html
}
\usage{
sample_big_data_annotated
}
\description{
Annotated/analyzed corpus of documents
}
\keyword{datasets}
