% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_haitian_fathers}
\alias{sample_haitian_fathers}
\title{Fragments of interviews about fatherhood}
\format{
\subsection{\code{sample_haitian_fathers}}{

A character vector with 6 elements
}
}
\source{
#' Zizi, 1996, p. 170, 221. (Haitian Father Data). Quoted and analyzed in Auerbach, C., & Silverstein, L. B. (2003). Qualitative data: an introduction to coding and analysis. New York University Press. p. 49-53
}
\usage{
sample_haitian_fathers
}
\description{
Annotated/analyzed in sample_haitian_fathers_annotated
}
\keyword{datasets}
