% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{start_gui}
\alias{start_gui}
\title{GUI for annotating / analyzing / codifying documents}
\usage{
start_gui(
  input = FALSE,
  docs = FALSE,
  codes = FALSE,
  show_monitor = FALSE,
  annotation_user = "default_user"
)
}
\arguments{
\item{input}{A list element from importing a .json file}

\item{docs}{Character vector with documents}

\item{codes}{Character vector with codes}

\item{show_monitor}{Logical show variables in GUI}

\item{annotation_user}{Character string with the annotators name}
}
\value{
false
}
\description{
Should open default browser.
Data MUST be exported as .json files!
You need to terminate the process after ending.
}
\examples{
\dontrun{
start_gui()
start_gui(input = mincaqdasr::sample_haitian_fathers_annotated)
start_gui(docs = mincaqdasr::sample_haitian_fathers, codes = c('role models', 'religious idea'))
}
}
