% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{zizi1996}
\alias{zizi1996}
\title{Returns a vector of sample documents}
\usage{
zizi1996()
}
\value{
A character vector.
}
\description{
Returns a vector of sample documents
}
\examples{
docs <- zizi1996()
}
